Create README.md
To read for me was very difficult, I was burn in Africa, to  read was not seample for me.  i was having a quiet read. After my inscription in the school I decided to lern english. The book was a good read. To be the child in school was the time i regrets, the story for my reading was dramatic. I don't know to read and now i decided to  read.